refactor: update import paths for ChordViewerNavigationMethods to use chord-viewer.types
